requirements.txt
django==2.2.1
pymysql
icecola
django-appconf==1.0.3
django-bootstrap-toolkit==2.15.0
django-compressor==2.2
django-filter==2.1.0
django-filters==0.2.1
django-jsonfield==1.0.1
django-markdown-deux==1.0.5
django-model-utils==3.1.2
django-redis==4.10.0
django-registration-redux==2.5
django-rest-swagger==2.2.0
django-simpleui==2.1
djangorestframework==3.9.1
Pillow==5.4.1